Error Fixes #15 - files were all updated to ensure functionality ~ ST10321423

+++ app/src/main/java/com/st10083866/prog7313_poe_personalbudgetapp/ui/home/MainPageActivity.kt
@@ -40,7 +40,7 @@ class MainPageActivity : AppCompatActivity() {
 
         //loads userID from saved session
         val session = SessionManager(this)
-        userId = intent.getIntExtra("USER_ID", session.getUserId().toInt())
+        userId = intent.getIntExtra("USER_ID", session.getUserId())
 
         //default fragment
         loadFragment(HomeFragment().withUser(userId))
